landing page bug fixes
