+++ frontend/src/componments/Button/Button.css
@@ -1,64 +1,64 @@
 .button {
-    border: none;
-    border-radius: 8px;
-    background: var(--Primary-Colour, #6A84C5);
-    
-    /* Other Panel Drop Shadow */
-    box-shadow: 0px 2px 14.8px -5px rgba(13, 41, 68, 0.12);
+  border: none;
+  border-radius: 8px;
+  background: var(--Primary-Colour, #6a84c5);
 
-    display: flex;
-    min-height: 32px;
-    padding: 11px 15px;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    gap: 10px;
+  /* Other Panel Drop Shadow */
+  box-shadow: 0px 2px 14.8px -5px rgba(13, 41, 68, 0.12);
 
-    color: var(--White-Colour, #FFF);
-    text-align: center;
-    font-family: Roboto;
-    font-size: 16px;
-    font-style: normal;
-    font-weight: 500;
-    line-height: 18px; /* 112.5% */
-    margin: auto;
-    margin-top: 16px;
-    margin-bottom: 16px;
-    cursor: pointer;
+  display: flex;
+  min-height: 32px;
+  padding: 11px 15px;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 10px;
+
+  color: var(--White-Colour, #fff);
+  text-align: center;
+  font-family: Roboto;
+  font-size: 16px;
+  font-style: normal;
+  font-weight: 500;
+  line-height: 18px; /* 112.5% */
+  /* margin: auto; */
+  margin-top: 16px;
+  margin-bottom: 16px;
+  cursor: pointer;
 }
 
 .button:hover {
-    border-radius: 8px;
-    background: var(--Darker-Primary-Colour, #32519E);
+  border-radius: 8px;
+  background: var(--Darker-Primary-Colour, #32519e);
 
-    /* Other Panel Drop Shadow */
-    box-shadow: 0px 2px 14.8px -5px rgba(13, 41, 68, 0.12);
+  /* Other Panel Drop Shadow */
+  box-shadow: 0px 2px 14.8px -5px rgba(13, 41, 68, 0.12);
 }
 
 .button-disabled {
-    border-radius: 8px;
-    border: none;
-    background: var(--Light-Prototype, #C2C2C2);
-    
-    /* Other Panel Drop Shadow */
-    box-shadow: 0px 2px 14.8px -5px rgba(13, 41, 68, 0.12);
+  border-radius: 8px;
+  border: none;
+  background: var(--Light-Prototype, #c2c2c2);
 
-    display: flex;
-    min-height: 32px;
-    padding: 15px 32px;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    gap: 10px;
+  /* Other Panel Drop Shadow */
+  box-shadow: 0px 2px 14.8px -5px rgba(13, 41, 68, 0.12);
 
-    color: var(--Grey-Text-Colour, #747474);
-    text-align: center;
-    font-family: Roboto;
-    font-size: 16px;
-    font-style: normal;
-    font-weight: 500;
-    line-height: 18px; /* 112.5% */
-    margin: auto;
-    margin-top: 16px;
-    margin-bottom: 16px;
+  display: flex;
+  min-height: 32px;
+  padding: 15px 32px;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 10px;
+
+  color: var(--Grey-Text-Colour, #747474);
+  text-align: center;
+  font-family: Roboto;
+  font-size: 16px;
+  font-style: normal;
+  font-weight: 500;
+  line-height: 18px; /* 112.5% */
+  margin: auto;
+  margin-top: 16px;
+  margin-bottom: 16px;
 }
