+++ frontend/src/componments/SignInButton/index.js
@@ -1,0 +1,1 @@
+export { default } from './SignInButton';
