+++ frontend/src/screens/SignIn.css
@@ -1,4 +1,6 @@
 .sign-in-button-wrapper {
-    display: grid;
-    justify-content: space-evenly;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: flex-start;
 }
